Corrected the route, removed the CreateVolunteer method from the controller, created a separate DTOs folder
