+++ src/Kind_heart.API/Program.cs
@@ -1,4 +1,5 @@
 using Kind_heart.API;
+using Kind_heart.Application.Volunteers;
 using Kind_heart.Application.Volunteers.CreateVolunteer;
 using Kind_heart.Infrastructure;
 using Kind_heart.Infrastructure.Repositories;
