+++ src/Kind_heart.Application/DTOs/RequisiteDto.cs
@@ -1,0 +1,3 @@
+namespace Kind_heart.Application.DTOs;
+
+public record RequisiteDto(string Name, string Description);
