+++ src/Kind_heart.Application/DTOs/SocialNetworkDto.cs
@@ -1,0 +1,3 @@
+namespace Kind_heart.Application.DTOs;
+
+public record SocialNetworkDto(string Name, string Path);
