+++ src/Kind_heart.Application/Volunteers/CreateVolunteer/CreateVolunteerRequest.cs
@@ -1,3 +1,4 @@
+using Kind_heart.Application.DTOs;
 using Kind_heart.Domain.ValueObjects;
 
 namespace Kind_heart.Application.Volunteers.CreateVolunteer;
@@ -11,6 +12,3 @@ public record CreateVolunteerRequest(string FirstName,
                                     string Phone,
                                     List<RequisiteDto> Requisites,
                                     List<SocialNetworkDto> SocialNetworks);
-
-public record RequisiteDto(string Name, string Description);
-public record SocialNetworkDto(string Name, string Path);
